update the links of question blocks
